Consolidate router test directories.
